requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.14
django-configurations==2.2
django-cors-headers==3.4.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
python-http-client==3.2.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sendgrid==6.4.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
whitenoise==5.1.0
